chore: quick renaming of a minor script for clarity
