+++ src/sports_calendar/core/selection/filters/definitions.py
